+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-07-15 15:24:16
-// Commit: feat(test): add bi-directional event bus handler for test
+// Updated at: 2026-07-15 16:01:17
+// Commit: chore(test): update dependencies and cleanup unused imports in test
 
 class testComponentState {
-  final String commitRef = '2854';
-  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
-  final DateTime timestamp = DateTime.parse('2026-07-15T15:24:16+05:00');
+  final String commitRef = '2857';
+  final String lastMessage = 'chore(test): update dependencies and cleanup unused imports in test';
+  final DateTime timestamp = DateTime.parse('2026-07-15T16:01:17+05:00');
 }